Depart Times
GridPane for depart times has been moved to allow the button to change the text in the text area. The times are currently hard-coded. We should make a class that has random outputs to make it look natural. Commented out fileIO part because it produced errors for me and isn't too helpful for now.